Create factory mode global variable5.

+++ wifi/services/wifi_standard/wifi_framework/wifi_manage/wifi_manager.cpp
@@ -103,7 +103,8 @@ int WifiManager::Init()
 int WifiManager::InitPart1()
 {
     int lastState = WifiConfigCenter::GetInstance().GetPersistWifiState(INSTID_WLAN0);
-    if (lastState != WIFI_STATE_DISABLED && WifiSettings::GetInstance().GetSystemMode() != SystemMode::FACTORY_MODE) {
+    if (lastState != WIFI_STATE_DISABLED 
+        && WifiSettings::GetInstance().GetSystemMode() != static_cast<int>(SystemMode::FACTORY_MODE)) {
         /* Automatic startup upon startup */
         WIFI_LOGI("AutoStartServiceThread lastState:%{public}d", lastState);
         WifiConfigCenter::GetInstance().SetWifiToggledState(lastState, INSTID_WLAN0);
